moved post video to external file
